requirements.txt
aiogram==3.4.1
openai>=1.12.0
psycopg2-binary==2.9.9
tavily-python==0.3.3
